+++ runelite-client/src/main/java/net/runelite/client/plugins/multicombatborders/MultiCombatBorder.java
@@ -22,7 +22,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package net.runelite.client.plugins.multilines;
+package net.runelite.client.plugins.multicombatborders;
 
 import lombok.Value;
 import net.runelite.api.coords.WorldPoint;
