+++ runelite-client/src/main/java/net/runelite/client/plugins/multicombatborders/MultiCombatBordersOverlay.java
@@ -22,7 +22,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package net.runelite.client.plugins.multilines;
+package net.runelite.client.plugins.multicombatborders;
 
 import java.awt.Dimension;
 import java.awt.Graphics2D;
